Removed the LoginLink class.  All other HeaderLinks are defined within the MainNav file, so its nice to be consistent and define this one there too.
