--- apps/site-landing/src/components/AppLinks/styles.module.less
@@ -1,6 +1,0 @@
-
-.loginLink{
-  text-decoration: uppercase;
-  border-bottom: none;
-  padding-bottom: 0;
-}
